backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
motor>=3.3.0
pydantic>=2.0.0
httpx>=0.24.0
python-dotenv>=1.0.0

